+++ lib/ncore/rails/active_model.rb
@@ -6,7 +6,6 @@ module NCore
       include ::ActiveModel::Conversion
       extend  ::ActiveModel::Naming
       extend  ::ActiveModel::Translation
-      alias :errors :errors_for_actionpack
     end
 
     if defined?(::Rails)
@@ -27,15 +26,6 @@ module NCore
       delete(*args)
     end
 
-    # actionpack 4 requires a more robust Errors object
-    def errors_for_actionpack
-      e0 = ::ActiveModel::Errors.new(self)
-      @errors.each do |e|
-        e0.add :base, e
-      end
-      e0
-    end
 
   end
-
 end